Refactor endpoint trace graph mapping ops
